moved search checkboxes to homepagecard
