+++ src/Navbar.jsx
@@ -6,7 +6,7 @@ export default function Navbar() {
     return (
         <div className="topnav">
             <a href="#register">Register/Login</a>
-            <a href="#search">Search</a>
+
             <a href="#latest">Latest</a>
             <a class="active" href="#home">Home</a>
             <div className="welcome">
